# Optimize code