Benchmark results:

Before:
```
BenchmarkGenerate-4                      5000000               311 ns/op               0 B/op          0 allocs/op
BenchmarkGenerateMaxSequence-4          20000000                66.9 ns/op             0 B/op          0 allocs/op
```
After:
```
BenchmarkGenerate-4                      5000000               243 ns/op               0 B/op          0 allocs/op
BenchmarkGenerateMaxSequence-4          20000000                66.6 ns/op             0 B/op          0 allocs/op